feat(core): per-field wire type for inbound message payload fields
An inbound `message<T>` payload field used to ride a uniform i32 wire (Q46
uniform lift). This truncated fractional `number` fields (0.8 → 0) and gave
boolean fields their type only by accident. The wire type is now decided per
field, driven from usage:

- Default wire is f32 (a declared `number` keeps its fraction).
- A `boolean` field seals to bool the moment it is consumed in a bool sink
  (`state.bool.write`, `buffer.bool.write`, `select` cond, `not`,
  `event.emitIf` cond, `midi.emitIf` cond) via the new
  `sealInboundFieldBool()` helper introduced in the previous commit.
- The declared `MessageDeclAst.fields[].wireType` is the single source of
  truth read by (a) the WASM emit path (`f32.load` vs `i32.load`), (b) the
  three transport encode sites (main SAB inject, worklet postMessage
  fallback, offline harness), and (c) the main-side `MessageGraphPayload<T>`
  handler type (`number` → `Node<'f32'>`).

Existing tests are updated for the new type surface (state.f32 where a
`number` field is written; `Node<'f32'>` type assertions; conformance SAB
simulation uses `Float32Array`). The `messageScalar` golden hash is
regenerated because the wire type change flips schemaHash / wasmSha /
snapshot hash. A v0.1.0 snapshot may fail to restore on v0.1.1; migration
is not provided because 0.1.x has no wild snapshot base.

The regression test that pins the fractional case (`messageGain` sends 0.8
and asserts 0.8 survives) is added in the next commit.

+++ packages/core/src/__tests__/behavior/typed-array-surface.test.ts
@@ -58,7 +58,7 @@ test("typed-array surface: element-type contract is verified by vp check typeche
   // Scalar message fields are lifted to Node on the handler side. Leaking them as raw
   // JS values would cause slot + 1 / if(armed) to run against a proxy at capture time
   // and corrupt the graph.
-  expectTrue<AssignableTo<SlotField, Node<"i32">>>();
+  expectTrue<AssignableTo<SlotField, Node<"f32">>>();
   expectFalse<AssignableTo<SlotField, number>>();
   expectTrue<AssignableTo<ArmedField, Node<"bool">>>();
   expectFalse<AssignableTo<ArmedField, boolean>>();
